requirements.txt
a==1.0
apptools==4.5.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.5
configobj==5.0.6
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
envisage==4.9.2
# filterpy==1.4.5
freetype-py==2.1.0.post1
future==0.18.2
glfw==1.11.2
gnureadline==8.0.0
imgui==1.1.0
importlib-metadata==1.6.0
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.2
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
msgpack-python==0.5.6
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
# numpy==1.18.4
open3d==0.9.0.0
packaging==20.3
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
prometheus-client==0.7.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pyface==7.0.0
Pygments==2.6.1
PyOpenGL==3.1.5
pyparsing==2.4.7
PyQt5==5.14.0
PyQt5-sip==12.7.2
pyrsistent==0.16.0
python-dateutil==2.8.1
# PyYAML==5.3.1
pyzmq==19.0.1
# scikit-learn==0.22.2.post1
# scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
terminado==0.8.3
testpath==0.4.4
# torch==1.5.0
torchdiffeq==0.1.0
# torchvision==0.6.0
tornado==6.0.4
# tqdm==4.48.2
traitlets==4.3.3
traits==6.0.0
traitsui==7.0.0
umsgpack==0.1.0
vispy==0.6.4
vtk==8.1.2
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.1.0
#MinkowskiEngine==0.5.4                                                          
PyYAML==6.0                                                                     
click==7.1.2                                                                    
easydict==1.9                                                                   
filterpy==1.4.5                                                                 
numba==0.49.1                                                                   
# numpy==1.19.5                                                                   
pytorch_lightning==1.5.9                                                        
scikit_learn==1.0.2 
scipy==1.5.0                                                                    
#spconv==1.2.1                                                                   
#torch==1.9.0                                                                    
torch_scatter==2.0.5
tqdm==4.48.2
protobuf==3.20.0
